Implementing scheduling for http client (#6)
* add scheduler for downloading jokes from API every indicated time

* add test for scheduler

+++ src/main/java/com/example/functionalbackend/joke/model/Joke.java
@@ -34,9 +34,5 @@ public class Joke {
     private TypeOfJoke type;
     private String joke;
     @Column(name = "id_of_joke")
-    private Integer id0fJoke;
-
-
-
-
+    private Long id0fJoke;
 }
